Added untested Bulletin insertion logic, speced out endorsements
The test coverage will not be complete until we can select from the db and
inspect the records that can be pulled out. Additionally a refactor of
protocol/ombproto looks nessecary so that the relationships between wire
records and their ombproto containers is more straight forward.

This could be accomplished by flipping to ombproto/ above procotocl/.

The wirerecord package may also become just wire or ombwire and live in its own
directory besides ombproto.